fix(ci): Fix failing TestDockerCompose by using curl for healthcheck
The `TestDockerCompose` integration test was failing in the CI pipeline
due to a timeout. The error message "Docker services did not become
healthy in time" indicated an issue with the service healthchecks.

The `mcpxy-server` service is built on a minimal `distroless` base
image that does not include the `wget` utility. The `docker-compose.yml`
file was configured to use `wget` for the `mcpxy-server` healthcheck,
which caused the healthcheck to fail and the service to never become
healthy.

This change fixes the issue by:
- Installing `curl` in the builder stage of the `mcpxy-server`
  Dockerfile and copying the `curl` binary to the final image.
- Updating the `docker-compose.yml` file to use `curl` for the
  `mcpxy-server` healthcheck.